<#88> Refactor: File, User, Follow 리팩터링

+++ src/main/java/com/linkerbell/portradebackend/domain/user/service/FollowService.java
@@ -4,11 +4,13 @@ import com.linkerbell.portradebackend.domain.user.domain.Follow;
 import com.linkerbell.portradebackend.domain.user.domain.User;
 import com.linkerbell.portradebackend.domain.user.dto.FollowersResponseDto;
 import com.linkerbell.portradebackend.domain.user.dto.FollowingsResponseDto;
-import com.linkerbell.portradebackend.domain.user.dto.ProfileResponeDto;
+import com.linkerbell.portradebackend.domain.user.dto.ProfileResponseDto;
 import com.linkerbell.portradebackend.domain.user.repository.FollowRepository;
 import com.linkerbell.portradebackend.domain.user.repository.UserRepository;
+import com.linkerbell.portradebackend.global.common.dto.PageResponseDto;
 import com.linkerbell.portradebackend.global.exception.ErrorCode;
 import com.linkerbell.portradebackend.global.exception.custom.NonExistentException;
+import com.linkerbell.portradebackend.global.exception.custom.UnAuthenticatedException;
 import com.linkerbell.portradebackend.global.exception.custom.UnAuthorizedException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -31,73 +33,88 @@ public class FollowService {
     private final FollowRepository followRepository;
 
     @Transactional
-    public void followUser(String follower, String following, User user) {
+    public boolean followUser(String follower, String following, User user) {
         User followUser = userRepository.findByUsername(following)
                 .orElseThrow(() -> new NonExistentException(ErrorCode.NONEXISTENT_USER));
+        userRepository.findByUsername(follower)
+                .orElseThrow(() -> new NonExistentException(ErrorCode.NONEXISTENT_USER));
 
-        if (!user.getUsername().equals(follower))
+        if (user == null) {
+            throw new UnAuthenticatedException(ErrorCode.NONEXISTENT_AUTHENTICATION);
+        }
+        if (!user.getUsername().equals(follower)) {
             throw new UnAuthorizedException(ErrorCode.NONEXISTENT_AUTHORIZATION);
+        }
 
         Optional<Follow> result = followRepository.findByFollowerIdAndFollowingId(follower, following);
-
-        // 팔로우 상태 체크 - 팔로우 된 상태
         if (result.isPresent()) {
             followRepository.delete(result.get());
+            return false;
         } else {
             Follow follow = Follow.builder()
                     .follower(user)
                     .following(followUser)
                     .build();
-
             followRepository.save(follow);
+            return true;
         }
     }
 
-    // 나를 팔로우 하는 user 조회
     public FollowersResponseDto getFollowers(String userId, int page, int size) {
+        userRepository.findByUsername(userId)
+                .orElseThrow(() -> new NonExistentException(ErrorCode.NONEXISTENT_USER));
+
         Pageable pageable = PageRequest.of(
                 page - 1,
                 size,
                 Sort.by(Sort.Direction.DESC, "id"));
+        Page<Follow> followPage = followRepository.findAllByFollowing_Username(pageable, userId);
 
-        Page<Follow> followUser = followRepository.findAllByFollowing_Username(pageable, userId);
-        List<ProfileResponeDto> followers = followUser
-                .stream()
-                .map(follow -> ProfileResponeDto.builder()
-                        .id(follow.getFollowerId())
+        List<ProfileResponseDto> followers = followPage.stream()
+                .map(follow -> ProfileResponseDto.builder()
+                        .id(follow.getFollowerUsername())
                         .name(follow.getFollowerName())
-                        .profileUrl(follow.getFollowerProfileUrl())
+                        .profileImageUrl(follow.getFollowerProfileUrl())
                         .job(follow.getFollowerJob())
                         .build())
                 .collect(Collectors.toList());
+        PageResponseDto pageResponseDto = PageResponseDto.builder()
+                .totalPage(followPage.getTotalPages())
+                .totalElement(followPage.getTotalElements())
+                .build();
 
         return FollowersResponseDto.builder()
+                .page(pageResponseDto)
                 .followers(followers)
-                .maxPage(followUser.getTotalPages())
                 .build();
     }
 
-    // 내가 팔로우 하는
     public FollowingsResponseDto getFollowings(String userId, int page, int size) {
+        userRepository.findByUsername(userId)
+                .orElseThrow(() -> new NonExistentException(ErrorCode.NONEXISTENT_USER));
+
         Pageable pageable = PageRequest.of(
                 page - 1,
                 size,
                 Sort.by(Sort.Direction.DESC, "id"));
+        Page<Follow> followPage = followRepository.findAllByFollower_Username(pageable, userId);
 
-        Page<Follow> followUser = followRepository.findAllByFollower_Username(pageable, userId);
-        List<ProfileResponeDto> followers = followUser
-                .stream()
-                .map(follow -> ProfileResponeDto.builder()
-                        .id(follow.getFollowingId())
+        List<ProfileResponseDto> followers = followPage.stream()
+                .map(follow -> ProfileResponseDto.builder()
+                        .id(follow.getFollowingUsername())
                         .name(follow.getFollowingName())
-                        .profileUrl(follow.getFollowingProfileUrl())
+                        .profileImageUrl(follow.getFollowingProfileUrl())
                         .job(follow.getFollowingJob())
                         .build())
                 .collect(Collectors.toList());
+        PageResponseDto pageResponseDto = PageResponseDto.builder()
+                .totalPage(followPage.getTotalPages())
+                .totalElement(followPage.getTotalElements())
+                .build();
 
         return FollowingsResponseDto.builder()
+                .page(pageResponseDto)
                 .followings(followers)
-                .maxPage(followUser.getTotalPages())
                 .build();
     }
 }
